requirements.txt
dj-database-url==0.5.0
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
django-rest-registration==0.6.3
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
gunicorn==20.1.0
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.9.1
PyJWT==2.0.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
whitenoise==5.2.0
